+++ client/src/main/java/org/openapitools/client/api/WelcomeApi.java
@@ -1,0 +1,114 @@
+package org.openapitools.client.api;
+
+import stroom.datasource.api.v2.*;
+import stroom.docref.*;
+import stroom.query.api.v2.*;
+
+import org.openapitools.client.model.*;
+
+
+
+import org.openapitools.client.ApiException;
+import org.openapitools.client.ApiClient;
+import org.openapitools.client.ApiResponse;
+import org.openapitools.client.Configuration;
+import org.openapitools.client.Pair;
+
+import javax.ws.rs.core.GenericType;
+
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class WelcomeApi {
+  private ApiClient apiClient;
+
+  public WelcomeApi() {
+    this(Configuration.getDefaultApiClient());
+  }
+
+  public WelcomeApi(ApiClient apiClient) {
+    this.apiClient = apiClient;
+  }
+
+  /**
+   * Get the API cilent
+   *
+   * @return API client
+   */
+  public ApiClient getApiClient() {
+    return apiClient;
+  }
+
+  /**
+   * Set the API cilent
+   *
+   * @param apiClient an instance of API client
+   */
+  public void setApiClient(ApiClient apiClient) {
+    this.apiClient = apiClient;
+  }
+
+  /**
+   * Get the configured HTML welcome message
+   * 
+   * @return Welcome
+   * @throws ApiException if fails to make API call
+   * @http.response.details
+     <table summary="Response Details" border="1">
+       <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
+       <tr><td> 0 </td><td> default response </td><td>  -  </td></tr>
+     </table>
+   */
+  public Welcome fetchWelcome() throws ApiException {
+    return fetchWelcomeWithHttpInfo().getData();
+  }
+
+  /**
+   * Get the configured HTML welcome message
+   * 
+   * @return ApiResponse&lt;Welcome&gt;
+   * @throws ApiException if fails to make API call
+   * @http.response.details
+     <table summary="Response Details" border="1">
+       <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
+       <tr><td> 0 </td><td> default response </td><td>  -  </td></tr>
+     </table>
+   */
+  public ApiResponse<Welcome> fetchWelcomeWithHttpInfo() throws ApiException {
+    Object localVarPostBody = null;
+    
+    // create path and map variables
+    String localVarPath = "/welcome/v1";
+
+    // query params
+    List<Pair> localVarQueryParams = new ArrayList<Pair>();
+    Map<String, String> localVarHeaderParams = new HashMap<String, String>();
+    Map<String, String> localVarCookieParams = new HashMap<String, String>();
+    Map<String, Object> localVarFormParams = new HashMap<String, Object>();
+
+
+    
+    
+    
+    final String[] localVarAccepts = {
+      "application/json"
+    };
+    final String localVarAccept = apiClient.selectHeaderAccept(localVarAccepts);
+
+    final String[] localVarContentTypes = {
+      
+    };
+    final String localVarContentType = apiClient.selectHeaderContentType(localVarContentTypes);
+
+    String[] localVarAuthNames = new String[] { "ApiKeyAuth" };
+
+    GenericType<Welcome> localVarReturnType = new GenericType<Welcome>() {};
+
+    return apiClient.invokeAPI("WelcomeApi.fetchWelcome", localVarPath, "GET", localVarQueryParams, localVarPostBody,
+                               localVarHeaderParams, localVarCookieParams, localVarFormParams, localVarAccept, localVarContentType,
+                               localVarAuthNames, localVarReturnType, false);
+  }
+}